Use command line args
