+++ scripts/SetPrice.js
@@ -1,13 +1,24 @@
 const oracleAbi = require("../abi/contracts/oracle/Oracle.sol/Oracle.json")
+const { Command } = require("commander")
 const hre = require("hardhat")
 const fs = require("fs")
+const program = new Command()
 
-module.exports.setPrice = async (privateKeyFile, price, oracleAddress) => {
-    // deploy all contracts using hardhat deploy
+program
+    .version("0.0.1")
+    .requiredOption("-pk, --privatekey", "private key filename")
+    .requiredOption("-p, --price", "price to set, in dollars")
+    .requiredOption("-o, --oracleAddress", "address of oracle to set price on")
+    .parse(process.argv)
+
+async function main() {
     const { ethers } = hre
+    const privateKeyFilename = program.args[0]
+    const price = parseInt(parseFloat(program.args[1]) * 100000000)
+    const oracleAddress = program.args[2]
 
     // Get private key from pkey2.secret, and trim any newlines
-    let pkey = fs.readFileSync(privateKeyFile, "utf-8")
+    let pkey = fs.readFileSync(privateKeyFilename, "utf-8")
     pkey = pkey.trim()
 
     // Create wallet with this private key, on kovan
@@ -17,3 +28,12 @@ module.exports.setPrice = async (privateKeyFile, price, oracleAddress) => {
     let oracle = new ethers.Contract(oracleAddress, oracleAbi)
     await oracle.connect(wallet).setPrice(price)
 }
+
+// We recommend this pattern to be able to use async/await everywhere
+// and properly handle errors.
+main()
+    .then(() => process.exit(0))
+    .catch((error) => {
+        console.error(error)
+        process.exit(1)
+    })
